requirements.txt
django==3.2
django-cors-headers==4.2.0
django-split-settings==1.2.0
psycopg2==2.9.6
uwsgi~=2.0.21
flake8==4.0
pytest~=7.3.2
python-dotenv==1.0.0